add `pangolin-runtime-system-params`
