+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
 	"bin/node/primitives",
 	"bin/node/runtime/pangolin",
 	"bin/node/runtime/pangolin/src/constants",
+	"bin/node/runtime/pangolin/src/pallets/system/params",
 	"client/cli",
 	"client/dvm/db",
 	"client/dvm/mapping-sync",
