+++ bin/node/runtime/pangolin/src/pallets/system/params/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+authors    = ["Darwinia Network <hello@darwinia.network>"]
+edition    = "2018"
+homepage   = "https://darwinia.network/"
+license    = "Unlicense"
+name       = "pangolin-runtime-system-params"
+readme     = "README.md"
+repository = "https://github.com/darwinia-network/darwinia-common/"
+version    = "2.3.0"
+
+[dependencies]
+static_assertions = { version = "1.1.0" }
+# substrate
+frame-support = { default-features = false, git = "https://github.com/darwinia-network/substrate.git", branch = "s2s" }
+frame-system  = { default-features = false, git = "https://github.com/darwinia-network/substrate.git", branch = "s2s" }
+sp-runtime    = { default-features = false, git = "https://github.com/darwinia-network/substrate.git", branch = "s2s" }
+
+[features]
+default = ["std"]
+
+std = [
+	"frame-support/std",
+	"frame-system/std",
+	"sp-runtime/std",
+]
